# improve perf for TableSharedKeyPipelinePolicy