--- a/sdk/tables/Azure.Data.Tables/src/TableSharedKeyPipelinePolicy.cs
+++ b/sdk/tables/Azure.Data.Tables/src/TableSharedKeyPipelinePolicy.cs
@@ -30,6 +30,9 @@ namespace Azure.Data.Tables
                 return ComputeSasSignature(credential, message);
             }
         }
+
+        private const string compQueryParam = "comp=";
+
         /// <summary>
         /// Shared key credentials used to sign requests
         /// </summary>
@@ -89,23 +92,17 @@ namespace Azure.Data.Tables
                 cr.Append('/');
             }
 
-            System.Collections.Generic.IDictionary<string, string> parameters = GetQueryParameters(resource); // Returns URL decoded values
-            if (parameters.Count > 0)
+            // If the request URI addresses a component of the resource, append the appropriate query string.
+            // The query string should include the question mark and the comp parameter (for example, ?comp=metadata).
+            // No other parameters should be included on the query string.
+            // https://docs.microsoft.com/en-us/rest/api/storageservices/authorize-with-shared-key#shared-key-lite-and-table-service-format-for-2009-09-19-and-later
+            if (TryGetCompQueryParameterValue(resource, out string compValue))
             {
-                foreach (var name in parameters.Keys.OrderBy(key => key, StringComparer.Ordinal))
-                {
-                    // If the request URI addresses a component of the resource, append the appropriate query string.
-                    // The query string should include the question mark and the comp parameter (for example, ?comp=metadata).
-                    // No other parameters should be included on the query string.
-                    // https://docs.microsoft.com/en-us/rest/api/storageservices/authorize-with-shared-key#shared-key-lite-and-table-service-format-for-2009-09-19-and-later
-                    if (name == "comp")
-                    {
 #pragma warning disable CA1308 // Normalize strings to uppercase
-                        cr.Append('?').Append(name.ToLowerInvariant()).Append('=').Append(parameters[name]);
+                cr.Append("?=").Append(compValue);
 #pragma warning restore CA1308 // Normalize strings to uppercase
-                    }
-                }
             }
+
             return cr.ToString();
         }
         public static IDictionary<string, string> GetQueryParameters(Uri uri)
@@ -133,6 +130,34 @@ namespace Azure.Data.Tables
                 }
             }
             return parameters;
+        }
+
+        public static bool TryGetCompQueryParameterValue(Uri uri, out string value)
+        {
+            value = null;
+
+            if (uri.Query == null || uri.Query.Length <= 0)
+            {
+                return false;
+            }
+            var query = uri.Query.AsSpan();
+            int iComp = query.IndexOf(compQueryParam.AsSpan(), StringComparison.OrdinalIgnoreCase) + compQueryParam.Length;
+            if (iComp < 0)
+            {
+                return false;
+            }
+
+            query = query.Slice(iComp);
+
+            int iEndOfValue = query.IndexOf('&');
+
+            value = iEndOfValue switch
+            {
+                -1 => query.Slice(0).ToString(),
+                _ => query.Slice(0, iEndOfValue).ToString()
+            };
+
+            return true;
         }
     }
 }