fixes #19672 

Note: input for *QueryFound is a sas token Uri, which is the worst case

```
|                       Method |        Mean |     Error |    StdDev |  Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------- |------------:|----------:|----------:|-------:|--------:|-------:|------:|------:|----------:|
| GetQueryParametersQueryFound | 2,287.79 ns | 29.147 ns | 24.339 ns | 179.02 |    2.46 | 0.4883 |     - |     - |    4105 B |
|         TryGetCompQueryFound |    99.12 ns |  1.183 ns |  1.106 ns |   7.77 |    0.11 | 0.0057 |     - |     - |      48 B |
|    GetQueryParametersNoQuery |    28.85 ns |  0.591 ns |  0.524 ns |   2.26 |    0.05 | 0.0095 |     - |     - |      80 B |
|            TryGetCompNoQuery |    12.76 ns |  0.135 ns |  0.126 ns |   1.00 |    0.00 |      - |     - |     - |         - |
```

![](https://media.giphy.com/media/26vUzl3ufQxx9dims/giphy.gif)